Fixed pidfile folder permissions
When configuring a custom pidfilepath on redhat the folder is not
created for mongod instances. This causes the pidfile to be created
in a different location (dbpath) and the service will be seen as
stopped by puppet. This problem is not present in mongos as it
sets DBPATH as the path of the pidfile. This change changes anyway
also mongos to be aligned with mongod pidfile path variable name.